requirements.txt
python-digitalocean
scaleway-sdk
